activity ,activityinfo page changes

+++ RedAlert.Entities/ComplexModels/DashboardResult.cs
@@ -9,10 +9,10 @@ namespace RedAlert.Entities.ComplexModels
 {
     public class DashboardResult
     {
-        public int Count_ActivitiesToComplete { get; set; }
-        public int Count_ActivitiesClosed { get; set; }
-        public int Count_UserActivitiesToComplete { get; set; }
-        public int Count_UserActivitiesClosed { get; set; }
+        public double Count_ActivitiesToComplete { get; set; }
+        public double Count_ActivitiesClosed { get; set; }
+        public double Count_UserActivitiesToComplete { get; set; }
+        public double Count_UserActivitiesClosed { get; set; }
     }
 
     public class DashboardResultParams : StoredProcedureParams
